Add blog post 'Analyzing Total Views on Velog'
A new blog post has been added to provide an in-depth guide on tracking total views for individual posts on Velog. The post covers the use of GraphQL API of Velog and explains how to sequentially use 'UserTags', 'Posts', and 'GetStats' queries to obtain the total views. This tutorial is created to aid those who want to track their blogs' performance on Velog.